backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
reportlab
python-multipart
transformers 
torch
vaderSentiment
numpy
scikit-learn
google-generativeai